+++ redesocial/src/main/java/br/org/serratec/redesocial/service/UsuarioService.java
@@ -13,6 +13,7 @@ import br.org.serratec.redesocial.domain.Usuario;
 import br.org.serratec.redesocial.dto.UsuarioDTO;
 import br.org.serratec.redesocial.dto.UsuarioInserirDTO;
 import br.org.serratec.redesocial.exception.EmailException;
+import br.org.serratec.redesocial.exception.NotFoundException;
 import br.org.serratec.redesocial.exception.SenhaException;
 import br.org.serratec.redesocial.repository.UsuarioRepository;
 import jakarta.transaction.Transactional;
@@ -31,11 +32,20 @@ public class UsuarioService {
 	
 	public UsuarioDTO buscar(Long id) {
 		Optional<Usuario> usuarioOpt = usuarioRepository.findById(id);
+		if(!usuarioOpt.isPresent()) {
+			throw new NotFoundException("Usuário não encontrado, ID: " + id);
+		}
 		return new UsuarioDTO(usuarioOpt.get());
 	}
 	
 	@Transactional
 	public UsuarioDTO inserir(UsuarioInserirDTO usuarioInserirDTO) throws EmailException, SenhaException {
+		if (!usuarioInserirDTO.getSenha().equals(usuarioInserirDTO.getConfirmaSenha())) {
+			throw new SenhaException("Senha e Confirma Senha não são iguais");
+		}
+		if(usuarioRepository.findByEmail(usuarioInserirDTO.getEmail())!=null) {
+			throw new EmailException("Email já existente");
+		}
 		
 		
 		Usuario usuario = new Usuario();
@@ -54,6 +64,9 @@ public class UsuarioService {
 	
 	@Transactional
 	public Usuario att(Usuario usuario, Long id) {
+		if(!usuarioRepository.existsById(id)) {
+			throw new NotFoundException("Usuário não encontrado, ID: " + id);
+		}
 		
 		usuario.setId(id);
 		usuario = usuarioRepository.save(usuario);
@@ -62,6 +75,9 @@ public class UsuarioService {
 	
 	@Transactional
 	public Integer del(Long id) {
+		if(!usuarioRepository.existsById(id)) {
+			throw new NotFoundException("Usuário não encontrado, ID: " + id);
+		}
 
 		usuarioRepository.deleteById(id);
 		return 1;
